Omega_Sovereign_Core/requirements_deck.txt
streamlit>=1.30.0
plotly>=5.18.0
numpy>=1.24.0
pywebview>=4.0
